Attempted to add Color Profile via User
Using SharedPreferences package, I tried to persist a number representing a color scheme for the site. This would be saved and then persist across screens. Currently a button is used to increment the number, but can be modified to be streamlined.

Current errors include unsuccessful persistence across screens due to routes being explicitly defined in the main program, which need a profile defined to work