Added Questions, Theme Fixer Upper, Skils
Added questions so that each type has 10 (except speaking) and added a Text to Speech plugin to do the Playback Game. When doing a game series, the theme would sometimes be the previous one instead of the updated one. This has been fix by calling for it in each page. Skills are now a mandatory part of each question, as well as a unique id for debugging. This id can probably be used to check which are correct/incorrect. Questions for the speaking game have to exclude topics about decimals, 4+ digits, and fractions. Overall clean up.